api/requirements.txt
aiohttp==4.0.0a1
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.2
appdirs==1.4.3
appnope==0.1.0
appscript==1.0.1
arlulaapi==1.4.0
arlulacore==1.0.1
asgiref==3.6.0
asn1crypto==0.24.0
astroid==2.2.5
astropy==3.1.2
atomicwrites==1.3.0
attrs==19.1.0
Automat==0.8.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.1
bitarray==0.8.3
bkcharts==0.2
bleach==3.1.0
bokeh==1.0.4
boto==2.49.0
Bottleneck==1.2.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==0.8.0
clyent==1.2.2
colorama==0.4.1
conda==22.9.0
conda-build==3.23.1
conda-package-handling==1.9.0
conda-verify==3.1.1
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.6.1
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.6
cytoolz==0.10.1
dask==1.1.4
decorator==4.4.0
defusedxml==0.5.0
distlib==0.3.0
distributed==1.26.0
Django==3.2.17
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
docutils==0.14
electrumX==1.13.0
entrypoints==0.3
et-xmlfile==1.0.1
eth-typing==2.2.1
eth-utils==1.8.4
fastcache==1.0.2
filelock==3.0.10
Flask==1.0.2
future==0.17.1
gevent==1.4.0
glob2==0.6
gmpy2==2.0.8
greenlet==0.4.15
grequests==0.6.0
h5py==2.9.0
heapdict==1.0.0
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==1.6.0
incremental==17.5.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.16
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
keyring==18.0.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
libarchive-c==2.8
lief==0.9.0
llvmlite==0.28.0
locket==0.2.0
lxml==4.3.2
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
more-itertools==6.0.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
mypy-extensions==0.4.3
navigator-updater==0.2.1
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
nltk==3.4
nose==1.3.7
notebook==5.7.8
numba==0.43.1
numexpr==2.6.9
numpy==1.16.2
numpydoc==0.8.0
olefile==0.46
openpyxl==2.6.1
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parsel==1.5.2
parso==0.3.4
partd==0.3.10
path.py==11.5.0
pathlib2==2.3.3
patsy==0.5.1
pep8==1.7.1
pexpect==4.6.0
pgeocode==0.3.0
pickleshare==0.7.5
Pillow==5.4.1
pipenv==2018.11.26
pkginfo==1.5.0.1
pluggy==0.9.0
ply==3.11
prometheus-client==0.6.0
prompt-toolkit==2.0.9
Protego==0.1.16
psutil==5.6.1
ptyprocess==0.6.0
py==1.8.0
py-ecc==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
PyDispatcher==2.0.5
pyevmasm==0.2.3
pyflakes==2.1.1
Pygments==2.3.1
PyHamcrest==2.0.0
pylint==2.3.1
pylru==1.2.0
pyodbc==4.0.26
pyOpenSSL==19.0.0
pyparsing==2.3.1
pyrsistent==0.14.11
PySocks==1.6.8
pytest==4.3.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-dateutil==2.8.0
python-dotenv==0.21.1
pytz==2018.9
PyWavelets==1.0.2
PyYAML==5.1
pyzmq==18.0.0
QtAwesome==0.5.7
qtconsole==4.4.3
QtPy==1.7.0
queuelib==1.5.0
rangehttpserver==1.2.0
requests==2.21.0
rope==0.11.0
ruamel-yaml==0.15.46
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.2.1
Scrapy==1.8.0
seaborn==0.9.0
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
snowballstemmer==1.2.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.8
Sphinx==1.8.5
sphinxcontrib-websupport==1.1.0
spyder==3.3.3
spyder-kernels==0.4.2
SQLAlchemy==1.3.1
sqlparse==0.3.0
statsmodels==0.9.0
svgwrite==1.1.6
sympy==1.3
tables==3.5.1
tblib==1.3.2
terminado==0.8.1
testpath==0.4.2
toml==0.10.2
tomlkit==0.5.3
toolz==0.9.0
tornado==6.0.2
tqdm==4.31.1
traitlets==4.3.2
Twisted==19.10.0
typing-extensions==4.4.0
unicodecsv==0.14.1
urllib3==1.24.1
virtualenv==20.0.17
virtualenv-clone==0.5.3
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.11.1
wurlitzer==1.0.2
xlrd==1.2.0
XlsxWriter==1.1.5
xlwings==0.15.4
xlwt==1.3.0
zict==0.1.4
zipp==3.1.0
zope.interface==4.7.1
